Move insert_new_report and insert_new_report_overview to external.database.
