+++ components/server/src/database/reports.py
@@ -1,16 +1,13 @@
 """Reports collection."""
 
-from typing import Any
-
 import pymongo
 from pymongo.database import Database
 
-from external.database import sessions
 from model.report import Report
 from server_utilities.functions import iso_timestamp
 from server_utilities.type import ReportId
 
-from .filters import DOES_EXIST, DOES_NOT_EXIST
+from .filters import DOES_NOT_EXIST
 
 
 # Sort order:
@@ -52,43 +49,3 @@ def latest_reports_overview(database: Database, max_iso_timestamp: str = "") -> 
 def report_exists(database: Database, report_uuid: ReportId):
     """Return whether a report with the specified report uuid exists."""
     return report_uuid in database.reports.distinct("report_uuid")
-
-
-def insert_new_report(database: Database, delta_description: str, uuids, *reports) -> dict[str, Any]:
-    """Insert one or more new reports in the reports collection."""
-    _prepare_documents_for_insertion(database, delta_description, reports, uuids, last=True)
-    report_uuids = [report["report_uuid"] for report in reports]
-    database.reports.update_many({"report_uuid": {"$in": report_uuids}, "last": DOES_EXIST}, {"$unset": {"last": ""}})
-    if len(reports) > 1:
-        database.reports.insert_many(reports, ordered=False)
-    else:
-        database.reports.insert_one(reports[0])
-    return dict(ok=True)
-
-
-def insert_new_reports_overview(database: Database, delta_description: str, reports_overview) -> dict[str, Any]:
-    """Insert a new reports overview in the reports overview collection."""
-    _prepare_documents_for_insertion(database, delta_description, [reports_overview])
-    database.reports_overviews.insert_one(reports_overview)
-    return dict(ok=True)
-
-
-def _prepare_documents_for_insertion(
-    database: Database, delta_description: str, documents, uuids=None, **extra_attributes
-) -> None:
-    """Prepare the documents for insertion in the database by removing any ids and setting the extra attributes."""
-    now = iso_timestamp()
-    user = sessions.user(database) or {}
-    email = user.get("email", "")
-    username = user.get("user", "An operator")
-    # Don't use str.format because there may be curly braces in the delta description, e.g. due to regular expressions:
-    description = delta_description.replace("{user}", username, 1)
-    for document in documents:
-        if "_id" in document:
-            del document["_id"]
-        document["timestamp"] = now
-        document["delta"] = dict(description=description, email=email)
-        if uuids:
-            document["delta"]["uuids"] = sorted(list(set(uuids)))
-        for key, value in extra_attributes.items():
-            document[key] = value
