+++ components/server/src/external/database/reports.py
@@ -1,15 +1,16 @@
 """Reports collection."""
 
-from typing import cast
+from typing import cast, Any
 
 import pymongo
 from pymongo.database import Database
 
 from database.filters import DOES_EXIST
 
-from server_utilities.functions import unique
+from server_utilities.functions import unique, iso_timestamp
 from server_utilities.type import MetricId, SubjectId
 
+from ..database import sessions
 from ..utils.type import Change
 
 
@@ -59,3 +60,43 @@ def metrics_of_subject(database: Database, subject_uuid: SubjectId) -> list[Metr
     projection: dict = {"_id": False, f"subjects.{subject_uuid}.metrics": True}
     report = database.reports.find_one(report_filter, projection=projection)
     return list(report["subjects"][subject_uuid]["metrics"].keys())
+
+
+def insert_new_report(database: Database, delta_description: str, uuids, *reports) -> dict[str, Any]:
+    """Insert one or more new reports in the reports collection."""
+    _prepare_documents_for_insertion(database, delta_description, reports, uuids, last=True)
+    report_uuids = [report["report_uuid"] for report in reports]
+    database.reports.update_many({"report_uuid": {"$in": report_uuids}, "last": DOES_EXIST}, {"$unset": {"last": ""}})
+    if len(reports) > 1:
+        database.reports.insert_many(reports, ordered=False)
+    else:
+        database.reports.insert_one(reports[0])
+    return dict(ok=True)
+
+
+def insert_new_reports_overview(database: Database, delta_description: str, reports_overview) -> dict[str, Any]:
+    """Insert a new reports overview in the reports overview collection."""
+    _prepare_documents_for_insertion(database, delta_description, [reports_overview])
+    database.reports_overviews.insert_one(reports_overview)
+    return dict(ok=True)
+
+
+def _prepare_documents_for_insertion(
+    database: Database, delta_description: str, documents, uuids=None, **extra_attributes
+) -> None:
+    """Prepare the documents for insertion in the database by removing any ids and setting the extra attributes."""
+    now = iso_timestamp()
+    user = sessions.user(database) or {}
+    email = user.get("email", "")
+    username = user.get("user", "An operator")
+    # Don't use str.format because there may be curly braces in the delta description, e.g. due to regular expressions:
+    description = delta_description.replace("{user}", username, 1)
+    for document in documents:
+        if "_id" in document:
+            del document["_id"]
+        document["timestamp"] = now
+        document["delta"] = dict(description=description, email=email)
+        if uuids:
+            document["delta"]["uuids"] = sorted(list(set(uuids)))
+        for key, value in extra_attributes.items():
+            document[key] = value
