+++ components/server/src/external/initialization/report.py
@@ -6,7 +6,8 @@ import pathlib
 
 from pymongo.database import Database
 
-from database.reports import insert_new_report, insert_new_reports_overview, latest_reports_overview, report_exists
+from database.reports import latest_reports_overview, report_exists
+from ..database.reports import insert_new_report, insert_new_reports_overview
 
 
 def initialize_reports_overview(database: Database) -> None:
