+++ components/server/src/external/routes/metric.py
@@ -7,12 +7,13 @@ from pymongo.database import Database
 
 from database.datamodels import latest_datamodel
 from database.measurements import insert_new_measurement, latest_measurement
-from database.reports import insert_new_report, latest_reports
+from database.reports import latest_reports
 from model.data import MetricData, SubjectData
 from model.metric import Metric
 from server_utilities.type import MetricId, SubjectId
 
 from ..database.datamodels import default_metric_attributes
+from ..database.reports import insert_new_report
 from ..model.actions import copy_metric, move_item
 from ..utils.functions import sanitize_html, uuid
