+++ components/server/src/external/routes/report.py
@@ -8,13 +8,14 @@ import requests
 from pymongo.database import Database
 
 from database.datamodels import latest_datamodel
-from database.reports import insert_new_report, latest_report, latest_reports
+from database.reports import latest_report, latest_reports
 from model.data import ReportData
 from model.report import Report
 from server_utilities.functions import iso_timestamp
 from server_utilities.type import ReportId
 
 from ..database.measurements import recent_measurements
+from ..database.reports import insert_new_report
 from ..initialization.secrets import EXPORT_FIELDS_KEYS_NAME
 from ..model.actions import copy_report
 from ..model.transformations import (
