+++ components/server/src/external/routes/source.py
@@ -6,11 +6,12 @@ import bottle
 from pymongo.database import Database
 
 from database.datamodels import latest_datamodel
-from database.reports import insert_new_report, latest_reports
+from database.reports import latest_reports
 from model.data import MetricData, SourceData
 from server_utilities.type import EditScope, MetricId, ReportId, SourceId, SubjectId
 
 from ..database.datamodels import default_source_parameters
+from ..database.reports import insert_new_report
 from ..model.actions import copy_source, move_item
 from ..model.queries import is_password_parameter
 from ..model.transformations import change_source_parameter
